update,delete type film
update,delete type film

+++ controllers/Film.js
@@ -1,3 +1,4 @@
+require('dotenv').config();
 const Film = require("./../models/film");
 const Type = require("./../models/type");
 const path = require('path');
@@ -7,6 +8,9 @@ var {validationResult} = require('express-validator');
 // get list film
 const getListFilm = async (req, res) => {
     const films = await Film.find({});
+    films.map(film => {
+        return  film.image = process.env.URL+film.image;
+    })
      return res.status(200).json({ 
         success: true,
         films
@@ -15,9 +19,14 @@ const getListFilm = async (req, res) => {
 
 // get film by id
 const getFilm = async (req, res) => {
-    // const errors = validationResult(req);
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        res.status(400).json({ errors: errors.array() });
+        return;
+    }
     const {id} = req.params;
     const film = await Film.findById(id);
+    film.image =process.env.URL+film.image;
     return res.status(200).json({         
         success: true,
         film });
@@ -44,6 +53,11 @@ const createFilm = async (req, res) => {
 
 // update film
 const updateFilm = async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() });
+      return;
+    }
     const {id} = req.params;
     const film = req.body;
     if(req.file){
@@ -58,6 +72,11 @@ const updateFilm = async (req, res) => {
 
 // delete film
 const deleteFilm = async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        res.status(400).json({ errors: errors.array() });
+        return;
+    }
     const {id} = req.params;
     await Film.findByIdAndDelete(id);
     return res.status(200).json({  
@@ -66,30 +85,37 @@ const deleteFilm = async (req, res) => {
     });
   };
 
-/***Film Type***/
+/***FOREIGN TYPES***/
 // add Type film
 const addTypeFilm = async (req, res) => {
     const {id} = req.params;
     const newType = new Type(req.body);
-    const film = await User.findById(id);
+    const film = await Film.findById(id);
     newType.films = film;
     await newType.save();
-    Film.types.push(newType._id);
-    await User.save();
+    film.types.push(newType._id);
+    await film.save();
     return res.status(201).json({
         success: true,
         message:"Created Type success!"})
-
 };
 
 // get Type film
 const getTypeFilm = async (req, res) => {
     const {id} = req.params;
-    const film = await User.findById(id).populate('types');
+    const film = await Film.findById(id).populate('types');
     const {types} = film;
+    let listTypes =[];
+    types.map(type => {
+        listTypes.push({
+            _id: type._id,
+            name: type.name,
+            description: type.description,
+        });
+    })
     return res.status(200).json({ 
         success: true,
-        types });
+        listTypes });
 }
